# api/gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: Tekton Hub
    description: HTTP services for managing Tekton Hub
    version: "1.0"
servers:
    - url: https://api.hub.tekton.dev
paths:
...
    /catalog/{catalogName}/error:
        get:
            tags:
                - catalog
            summary: CatalogError catalog
            description: List all errors occurred refreshing a catalog
            operationId: catalog#CatalogError
            parameters:
                - name: catalogName
                  in: path
                  description: Name of catalog
                  required: true
                  schema:
                    type: string
                    description: Name of catalog
                    example: tekton
                  example: tekton
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CatalogErrorResponseBody'
                            example:
                                data:
                                    - errors:
                                        - Resource tekton.dev/v1beta1, Kind=Task - buildah has no display name
                                        - Resource tekton.dev/v1beta1, Kind=task - curl has no display name
                                      type: warning
                                    - errors:
                                        - Resource tekton.dev/v1beta1, Kind=Task - buildah has no display name
                                        - Resource tekton.dev/v1beta1, Kind=task - curl has no display name
                                      type: warning
                "500":
                    description: 'internal-error: Internal Server Error'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
            security:
                - jwt_header_Authorization:
                    - rating:read
                    - rating:write
                    - agent:create
                    - catalog:refresh
                    - config:refresh
                    - refresh:token
    /catalog/{catalogName}/refresh:
        post:
            tags:
                - catalog
            summary: Refresh catalog
            description: Refresh a Catalog by it's name
            operationId: catalog#Refresh
            parameters:
                - name: catalogName
                  in: path
                  description: Name of catalog
                  required: true
                  schema:
                    type: string
                    description: Name of catalog
                    example: tekton
                  example: tekton
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Job'
                            example:
                                catalogName: tekton
                                id: 1
                                status: queued
                "404":
                    description: 'not-found: Resource Not Found Error'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
                "500":
                    description: 'internal-error: Internal Server Error'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
            security:
                - jwt_header_Authorization:
                    - rating:read
                    - rating:write
                    - agent:create
                    - catalog:refresh
                    - config:refresh
                    - refresh:token
    /catalog/refresh:
        post:
            tags:
                - catalog
            summary: RefreshAll catalog
            description: Refresh all catalogs
            operationId: catalog#RefreshAll
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Job'
                                example:
                                    - catalogName: tekton
                                      id: 1
                                      status: queued
                                    - catalogName: tekton
                                      id: 1
                                      status: queued
                            example:
                                - catalogName: tekton
                                  id: 1
                                  status: queued
                                - catalogName: tekton
                                  id: 1
                                  status: queued
                                - catalogName: tekton
                                  id: 1
                                  status: queued
                                - catalogName: tekton
                                  id: 1
                                  status: queued
                "500":
                    description: 'internal-error: Internal Server Error'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
            security:
                - jwt_header_Authorization:
                    - rating:read
                    - rating:write
                    - agent:create
                    - catalog:refresh
                    - config:refresh
                    - refresh:token
...
components:
    schemas:
...
        CatalogErrorResponseBody:
            type: object
            properties:
                data:
                    type: array
                    items:
                        $ref: '#/components/schemas/CatalogErrors'
                    description: Catalog errors
                    example:
                        - errors:
                            - Resource tekton.dev/v1beta1, Kind=Task - buildah has no display name
                            - Resource tekton.dev/v1beta1, Kind=task - curl has no display name
                          type: warning
                        - errors:
                            - Resource tekton.dev/v1beta1, Kind=Task - buildah has no display name
                            - Resource tekton.dev/v1beta1, Kind=task - curl has no display name
                          type: warning
            example:
                data:
                    - errors:
                        - Resource tekton.dev/v1beta1, Kind=Task - buildah has no display name
                        - Resource tekton.dev/v1beta1, Kind=task - curl has no display name
                      type: warning
                    - errors:
                        - Resource tekton.dev/v1beta1, Kind=Task - buildah has no display name
                        - Resource tekton.dev/v1beta1, Kind=task - curl has no display name
                      type: warning
                    - errors:
                        - Resource tekton.dev/v1beta1, Kind=Task - buildah has no display name
                        - Resource tekton.dev/v1beta1, Kind=task - curl has no display name
                      type: warning
            required:
                - data
        CatalogErrors:
            type: object
            properties:
                errors:
                    type: array
                    items:
                        type: string
                        example: Aspernatur eos dolores adipisci vel.
                    description: Catalog Error message
                    example:
                        - Resource tekton.dev/v1beta1, Kind=Task - buildah has no display name
                        - Resource tekton.dev/v1beta1, Kind=task - curl has no display name
                type:
                    type: string
                    description: Catalog Errror type
                    example: warning
            description: CatalogErrors define the errors that occurred during catalog refresh
            example:
                errors:
                    - Resource tekton.dev/v1beta1, Kind=Task - buildah has no display name
                    - Resource tekton.dev/v1beta1, Kind=task - curl has no display name
                type: warning
            required:
                - type
                - errors
...
        Error:
            type: object
            properties:
                fault:
                    type: boolean
                    description: Is the error a server-side fault?
                    example: true
                id:
                    type: string
                    description: ID is a unique identifier for this particular occurrence of the problem.
                    example: 123abc
                message:
                    type: string
                    description: Message is a human-readable explanation specific to this occurrence of the problem.
                    example: parameter 'p' must be an integer
                name:
                    type: string
                    description: Name is the name of this class of errors.
                    example: bad_request
                temporary:
                    type: boolean
                    description: Is the error temporary?
                    example: false
                timeout:
                    type: boolean
                    description: Is the error a timeout?
                    example: false
            description: Invalid request body
            example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
            required:
                - name
                - id
                - message
                - temporary
                - timeout
                - fault
...
        Job:
            type: object
            properties:
                catalogName:
                    type: string
                    description: Name of the catalog
                    example: tekton
                id:
                    type: integer
                    description: id of the job
                    example: 1
                status:
                    type: string
                    description: status of the job
                    example: queued
            example:
                catalogName: tekton
                id: 1
                status: queued
            required:
                - id
                - catalogName
                - status
...
    securitySchemes:
        jwt_header_Authorization:
            type: http
            description: Secures endpoint by requiring a valid JWT retrieved via the /auth/login endpoint.
            scheme: bearer
...